+++ NEWS.md
@@ -1,5 +1,9 @@
 # testthat 0.11.0.9000
 
+* `throws_error` and `expect_error` can now take an `expected_class` argument to express
+   an expected subclass for the error condition (@dchudz)
+
+* Fix failure message for `throws_error` in case where no error is raised (@nealrichardson, #342).
 * `expect_output(f(), NA)` will fail if `f()` produces output; 
   `expect_output(f(), NULL)` will fail if `f()` doesn't produce output (#323).